+++ src/router/index.ts
@@ -11,8 +11,8 @@ import { commonUtil, translate } from '@common'
 import { useUserStore } from '@/store/user'
 
 import 'vue-router'
-import { useAuth } from '@/composables/useAuth';
-import Login from '@/views/Login.vue';
+import { useAuth } from '@common/composables/auth';
+import Login from '@common/components/Login.vue';
 
 
 const authGuard = async (to: any, from: any, next: any) => {
